Test cases added for intel XPU attention backend
